AdviceGen/Experimentation/requirements.txt
pydantic-settings
pydantic
openai
azure-identity